chore: renamed RequestController and changed routes
